src/agents/sow_evaluator.py
from typing import Dict, Any, List
import yaml
from pathlib import Path
import time


def evaluate_sow_document(content: str) -> Dict[str, Any]:
    """
    Placeholder SOW evaluator - Framework for future implementation.
    
    This is a simplified version that returns mock results.
    The actual evaluation logic will be implemented later.
    """
    
    # Simulate processing time
    time.sleep(2)
    
    # Mock evaluation results for demonstration
    phases = ["scope_definition", "dependencies_analysis", "assumptions_review"]
    
    # Create mock phase results
    phase_results = {}
    for phase in phases:
        phase_results[phase] = {
            "score": 2,  # Mock score out of 3
            "feedback": f"This is a placeholder evaluation for {phase.replace('_', ' ')}. Actual evaluation logic will be implemented later.",
            "strengths": [
                f"Framework ready for {phase.replace('_', ' ')} evaluation",
                "Extensible architecture in place"
            ],
            "gaps": [
                "Evaluation logic not yet implemented",
                "Criteria-specific analysis pending"
            ]
        }
    
    # Calculate mock overall score
    total_score = sum(r["score"] for r in phase_results.values())
    max_score = len(phases) * 3
    
    return {
        "evaluation_type": "statement_of_work",
        "overall_score": total_score,
        "max_score": max_score,
        "phase_results": phase_results,
        "key_findings": [
            "SOW evaluation framework is ready",
            "Placeholder results generated successfully",
            "Ready for actual evaluation logic implementation"
        ],
        "recommendations": [
            {
                "title": "Implement SOW Evaluation Logic",
                "description": "Add specific evaluation criteria and scoring logic for SOW documents",
                "priority": "High"
            },
            {
                "title": "Define SOW Quality Metrics",
                "description": "Establish clear metrics for scope, dependencies, and assumptions evaluation",
                "priority": "Medium"
            }
        ],
        "summary": f"SOW evaluation framework completed. Mock score: {total_score}/{max_score}"
    } 