requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.2.0
langchain-community>=0.2.0
pydantic>=2.8.0
python-dotenv>=1.0.0
python-docx>=0.8.11
PyPDF2>=3.0.1
pyyaml>=6.0
python-multipart>=0.0.6
watchdog>=3.0.0
typing-extensions>=4.12.0
plotly>=5.17.0 